new bash example: prepare data for CERA publication
The new script (and some input files and other scripts, which it uses)
takes CMAQ output data and brings them into a format for the
publication at CERA. The grid is not interpolated but properly
documented according to CF-conventions. Generally, the file
conforms with the CF-conventions. There are some standard names
with ambiguous meaning. Please read the READMe of the CERA data set
for details.